feat: 징계 회원 조회 VO 작성

+++ src/main/java/com/example/fieldpasserbe/admin/vo/PunishVO.java
@@ -1,6 +1,6 @@
 package com.example.fieldpasserbe.admin.vo;
 
-import com.example.fieldpasserbe.admin.dto.PunishResponceDTO;
+import com.example.fieldpasserbe.support.dto.PunishResponceDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,6 +14,6 @@ import java.util.List;
 @Builder
 public class PunishVO {
     private String resultCode;
-    private int resultDataNum;
+    private Long resultDataNum;
     private List<PunishResponceDTO> resultData;
 }
